refactor: seed image domain package
