+++ pkg/agentcompose/image_auto_backend_test.go
@@ -16,25 +16,25 @@ func TestAutoImageBackendUsesDockerWhenAutoPingSucceeds(t *testing.T) {
 	dockerCalled := false
 	ociCalled := false
 	backend := &AutoImageBackend{
-		mode: appconfig.ImageStoreModeAuto,
-		docker: &fakeImageBackend{listImages: func(ctx context.Context, req ImageListRequest) (ImageListResult, error) {
+		Mode: appconfig.ImageStoreModeAuto,
+		Docker: &fakeImageBackend{listImages: func(ctx context.Context, req ImageListRequest) (ImageListResult, error) {
 			dockerCalled = true
 			return ImageListResult{StoreStatus: &agentcomposev2.ImageStoreStatus{Store: agentcomposev2.ImageStoreKind_IMAGE_STORE_KIND_DOCKER_DAEMON}}, nil
 		}},
-		oci: &fakeImageBackend{listImages: func(ctx context.Context, req ImageListRequest) (ImageListResult, error) {
+		OCI: &fakeImageBackend{listImages: func(ctx context.Context, req ImageListRequest) (ImageListResult, error) {
 			ociCalled = true
 			return ImageListResult{}, nil
 		}},
-		pingDocker:  func(ctx context.Context) error { return nil },
-		pingTimeout: time.Second,
+		PingDocker:  func(ctx context.Context) error { return nil },
+		PingTimeout: time.Second,
 	}
 
 	result, err := backend.ListImages(context.Background(), ImageListRequest{})
 	if err != nil {
 		t.Fatalf("ListImages returned error: %v", err)
 	}
-	if !dockerCalled || ociCalled || backend.lastSelection != appconfig.ImageStoreModeDocker || result.StoreStatus.GetStore() != agentcomposev2.ImageStoreKind_IMAGE_STORE_KIND_DOCKER_DAEMON {
-		t.Fatalf("selection docker=%v oci=%v last=%q result=%#v", dockerCalled, ociCalled, backend.lastSelection, result)
+	if !dockerCalled || ociCalled || backend.LastSelection != appconfig.ImageStoreModeDocker || result.StoreStatus.GetStore() != agentcomposev2.ImageStoreKind_IMAGE_STORE_KIND_DOCKER_DAEMON {
+		t.Fatalf("selection docker=%v oci=%v last=%q result=%#v", dockerCalled, ociCalled, backend.LastSelection, result)
 	}
 }
 
@@ -42,25 +42,25 @@ func TestAutoImageBackendUsesOCIWhenAutoPingFails(t *testing.T) {
 	dockerCalled := false
 	ociCalled := false
 	backend := &AutoImageBackend{
-		mode: appconfig.ImageStoreModeAuto,
-		docker: &fakeImageBackend{pullImage: func(ctx context.Context, req ImagePullRequest) (ImagePullResult, error) {
+		Mode: appconfig.ImageStoreModeAuto,
+		Docker: &fakeImageBackend{pullImage: func(ctx context.Context, req ImagePullRequest) (ImagePullResult, error) {
 			dockerCalled = true
 			return ImagePullResult{}, nil
 		}},
-		oci: &fakeImageBackend{pullImage: func(ctx context.Context, req ImagePullRequest) (ImagePullResult, error) {
+		OCI: &fakeImageBackend{pullImage: func(ctx context.Context, req ImagePullRequest) (ImagePullResult, error) {
 			ociCalled = true
 			return ImagePullResult{ResolvedRef: "oci"}, nil
 		}},
-		pingDocker:  func(ctx context.Context) error { return errors.New("docker unavailable") },
-		pingTimeout: time.Second,
+		PingDocker:  func(ctx context.Context) error { return errors.New("docker unavailable") },
+		PingTimeout: time.Second,
 	}
 
 	result, err := backend.PullImage(context.Background(), ImagePullRequest{ImageRef: "team/app:latest"})
 	if err != nil {
 		t.Fatalf("PullImage returned error: %v", err)
 	}
-	if dockerCalled || !ociCalled || backend.lastSelection != appconfig.ImageStoreModeOCI || result.ResolvedRef != "oci" {
-		t.Fatalf("selection docker=%v oci=%v last=%q result=%#v", dockerCalled, ociCalled, backend.lastSelection, result)
+	if dockerCalled || !ociCalled || backend.LastSelection != appconfig.ImageStoreModeOCI || result.ResolvedRef != "oci" {
+		t.Fatalf("selection docker=%v oci=%v last=%q result=%#v", dockerCalled, ociCalled, backend.LastSelection, result)
 	}
 }
 
@@ -95,20 +95,20 @@ func TestAutoImageBackendForcedModesDoNotPing(t *testing.T) {
 			dockerCalled := false
 			ociCalled := false
 			backend := &AutoImageBackend{
-				mode: tc.mode,
-				docker: &fakeImageBackend{
+				Mode: tc.mode,
+				Docker: &fakeImageBackend{
 					inspectImage: func(ctx context.Context, req ImageInspectRequest) (ImageInspectResult, error) {
 						dockerCalled = true
 						return ImageInspectResult{}, nil
 					},
 				},
-				oci: &fakeImageBackend{
+				OCI: &fakeImageBackend{
 					removeImage: func(ctx context.Context, req ImageRemoveRequest) (ImageRemoveResult, error) {
 						ociCalled = true
 						return ImageRemoveResult{}, nil
 					},
 				},
-				pingDocker: func(ctx context.Context) error {
+				PingDocker: func(ctx context.Context) error {
 					pinged = true
 					return nil
 				},
@@ -116,8 +116,8 @@ func TestAutoImageBackendForcedModesDoNotPing(t *testing.T) {
 			if err := tc.run(backend); err != nil {
 				t.Fatalf("operation returned error: %v", err)
 			}
-			if pinged || backend.lastSelection != tc.want {
-				t.Fatalf("pinged=%v last=%q want=%q", pinged, backend.lastSelection, tc.want)
+			if pinged || backend.LastSelection != tc.want {
+				t.Fatalf("pinged=%v last=%q want=%q", pinged, backend.LastSelection, tc.want)
 			}
 			if tc.want == appconfig.ImageStoreModeDocker && !dockerCalled {
 				t.Fatalf("docker backend was not called")
